+++ packages/orq-rc/src/orq_ai_sdk/models/createcontactop.py
@@ -140,7 +140,7 @@ class CreateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-08-12T08:36:02.813Z")
+    updated: Optional[datetime] = parse_datetime("2025-08-12T12:59:35.165Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")