+++ packages/orq-rc/src/orq_ai_sdk/models/getonepromptop.py
@@ -45,6 +45,17 @@ r"""The modality of the model"""
 GetOnePromptFormat = Literal["url", "b64_json", "text", "json_object"]
 r"""Only supported on `image` models."""
 
+GetOnePromptResponseFormatPromptsResponseType = Literal["text"]
+
+
+class GetOnePromptResponseFormat3TypedDict(TypedDict):
+    type: GetOnePromptResponseFormatPromptsResponseType
+
+
+class GetOnePromptResponseFormat3(BaseModel):
+    type: GetOnePromptResponseFormatPromptsResponseType
+
+
 GetOnePromptResponseFormatPromptsType = Literal["json_object"]
 
 
@@ -86,7 +97,11 @@ class GetOnePromptResponseFormat1(BaseModel):
 
 GetOnePromptResponseFormatTypedDict = TypeAliasType(
     "GetOnePromptResponseFormatTypedDict",
-    Union[GetOnePromptResponseFormat2TypedDict, GetOnePromptResponseFormat1TypedDict],
+    Union[
+        GetOnePromptResponseFormat2TypedDict,
+        GetOnePromptResponseFormat3TypedDict,
+        GetOnePromptResponseFormat1TypedDict,
+    ],
 )
 r"""An object specifying the format that the model must output.
 
@@ -100,7 +115,11 @@ Important: when using JSON mode, you must also instruct the model to produce JSO
 
 GetOnePromptResponseFormat = TypeAliasType(
     "GetOnePromptResponseFormat",
-    Union[GetOnePromptResponseFormat2, GetOnePromptResponseFormat1],
+    Union[
+        GetOnePromptResponseFormat2,
+        GetOnePromptResponseFormat3,
+        GetOnePromptResponseFormat1,
+    ],
 )
 r"""An object specifying the format that the model must output.