+++ packages/orq-rc/src/orq_ai_sdk/models/getpromptversionop.py
@@ -73,6 +73,17 @@ r"""The modality of the model"""
 GetPromptVersionFormat = Literal["url", "b64_json", "text", "json_object"]
 r"""Only supported on `image` models."""
 
+GetPromptVersionResponseFormatPromptsResponseType = Literal["text"]
+
+
+class GetPromptVersionResponseFormat3TypedDict(TypedDict):
+    type: GetPromptVersionResponseFormatPromptsResponseType
+
+
+class GetPromptVersionResponseFormat3(BaseModel):
+    type: GetPromptVersionResponseFormatPromptsResponseType
+
+
 GetPromptVersionResponseFormatPromptsType = Literal["json_object"]
 
 
@@ -116,6 +127,7 @@ GetPromptVersionResponseFormatTypedDict = TypeAliasType(
     "GetPromptVersionResponseFormatTypedDict",
     Union[
         GetPromptVersionResponseFormat2TypedDict,
+        GetPromptVersionResponseFormat3TypedDict,
         GetPromptVersionResponseFormat1TypedDict,
     ],
 )
@@ -131,7 +143,11 @@ Important: when using JSON mode, you must also instruct the model to produce JSO
 
 GetPromptVersionResponseFormat = TypeAliasType(
     "GetPromptVersionResponseFormat",
-    Union[GetPromptVersionResponseFormat2, GetPromptVersionResponseFormat1],
+    Union[
+        GetPromptVersionResponseFormat2,
+        GetPromptVersionResponseFormat3,
+        GetPromptVersionResponseFormat1,
+    ],
 )
 r"""An object specifying the format that the model must output.