+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -100,7 +100,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K2EPRDBD5BG06E7RBR4A9JAY"
+        "01K2F5TZFA15V7HGQVNJHHSYCE"
     )
     r"""The id of the resource"""